+++ samples/apps/logging/logging.cpp
@@ -21,6 +21,7 @@
 #include "ccf/json_handler.h"
 #include "ccf/network_identity_interface.h"
 #include "ccf/version.h"
+#include "crypto/public_key.h"
 
 #include <charconv>
 #include <crypto/cose.h>
@@ -2145,6 +2146,40 @@ namespace loggingapp
           get_cose_endorsements, context, is_tx_committed),
         auth_policies)
         .set_auto_schema<void, LoggingGetCoseEndorsements::Out>()
+        .set_forwarding_required(ccf::endpoints::ForwardingRequired::Never)
+        .install();
+
+      auto get_trusted_keys = [&](
+                                ccf::endpoints::ReadOnlyEndpointContext& ctx) {
+        auto network_identity_subsystem =
+          context.get_subsystem<ccf::NetworkIdentitySubsystemInterface>();
+        if (network_identity_subsystem == nullptr)
+        {
+          ctx.rpc_ctx->set_error(
+            HTTP_STATUS_INTERNAL_SERVER_ERROR,
+            ccf::errors::InternalError,
+            "Network identity subsystem not available");
+          return;
+        }
+
+        auto keys = network_identity_subsystem->get_trusted_keys();
+        nlohmann::json jwks = nlohmann::json::object();
+        auto keys_array = nlohmann::json::array();
+        for (const auto& [seqno, key_ptr] : keys)
+        {
+          const auto kid = ccf::crypto::kid_from_key(key_ptr->public_key_der());
+          keys_array.push_back(key_ptr->public_key_jwk(kid));
+        }
+        jwks["keys"] = keys_array;
+
+        ctx.rpc_ctx->set_response_json(jwks, HTTP_STATUS_OK);
+      };
+      make_read_only_endpoint(
+        "/log/public/trusted_keys",
+        HTTP_GET,
+        get_trusted_keys,
+        ccf::no_auth_required)
+        .set_auto_schema<void, void>()
         .set_forwarding_required(ccf::endpoints::ForwardingRequired::Never)
         .install();